requirements.txt
# Core Dependencies
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
datasets==3.3.2

# NLP Evaluation Models
sentence-transformers==3.4.1
openai==1.64.0
accelerate==1.4.0

# Bias Mitigation (BBQ, Fairness Metrics)
fairlearn==0.8.0
seaborn==0.12.2

# Deception & Truthfulness Detection (TruthfulQA, Contradiction Models)
sentencepiece==0.2.0  # Required for some NLP models
nltk==3.9.1
spacy==3.8.4
scipy==1.15.2

# Miscellaneous Utilities
tqdm==4.66.3  # Progress bar
matplotlib==3.10.0  # Visualization
ipython==8.12.3  # Interactive debugging
jupyter==1.0.0  # For exploratory analysis

# # If using GPU
# cupy-cuda12x==12.0.0  # Only needed for CUDA acceleration

requests~=2.32.3
python-dotenv~=1.0.1
textattack~=0.3.10